+++ packages/shadow-objects/CHANGELOG.md
@@ -5,10 +5,14 @@ All notable changes to [@spearwolf/shadow-objects](https://github.com/spearwolf/
 The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- renamed `useResource()` to `createResource()` in `ShadowObjectParams` interface
+
 ## [0.23.0] - 2025-11-26
 
 - enhance the shadow-objects creation api _aka_ `ShadowObjectParams`
   - added the `useProperties()` function
-  - added the `useResource()` function
+  - added the `createResource()` function
 - added lots of new tests and improved code coverage